+++ _evan/lib/ev_gui/menu.cpp
@@ -1,11 +1,23 @@
 /* 
- * Copyright (C) 2021 Evan Pernu - All Rights Reserved
+ * Copyright (C) 2021 Evan Pernu. Author: Evan Pernu
+ * 
  * You may use, distribute and modify this code under the
  * terms of the GNU AGPLv3 license.
  * 
- * You should have received a copy of the GNU AGPLv3 license with
- * this file (LICENSE.md). If not, please write to: evanpernu@gmail.com, 
- * or visit: https://www.gnu.org/licenses/agpl-3.0.en.html
+ * This program is part of "Evan's Daisy Projects".
+ * 
+ * "Evan's Daisy Projects" is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 #include "menu.h"
@@ -28,6 +40,19 @@ void Menu::append(
         std::function<void ()> theCallback
 ){
     MenuItem* m = new MenuItem(theTitle, thePad, theValues, theDefault, theCallback);
+    items.push_back(m);
+    itemNameToIndex[m->getTitle()] = static_cast<uint8_t>(items.size()-1);
+}
+
+void Menu::append(
+        std::string theTitle,
+        std::string thePad,
+        int16_t theMin,
+        int16_t theMax,
+        int16_t theDefault,
+        std::function<void ()> theCallback
+){
+    MenuItem* m = new MenuItemInt(theTitle, thePad, theMin, theMax, theDefault, theCallback);
     items.push_back(m);
     itemNameToIndex[m->getTitle()] = static_cast<uint8_t>(items.size()-1);
 }