+++ _evan/lib/ev_theory/arp.h
@@ -1,11 +1,23 @@
 /* 
- * Copyright (C) 2021 Evan Pernu - All Rights Reserved
+ * Copyright (C) 2021 Evan Pernu. Author: Evan Pernu
+ * 
  * You may use, distribute and modify this code under the
  * terms of the GNU AGPLv3 license.
  * 
- * You should have received a copy of the GNU AGPLv3 license with
- * this file (LICENSE.md). If not, please write to: evanpernu@gmail.com, 
- * or visit: https://www.gnu.org/licenses/agpl-3.0.en.html
+ * This program is part of "Evan's Daisy Projects".
+ * 
+ * "Evan's Daisy Projects" is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 #include <string>
@@ -13,7 +25,7 @@
 
 #include "chord.h"
 
-namespace jellybeans {
+namespace ev_theory {
     class Arp {
     private:
         // Maximum arp steps
@@ -87,4 +99,4 @@ namespace jellybeans {
         "U+D Ex",
         "Random"
     };
-}
+} // namespace ev_theory