+++ _evan/Jellybeans/CHANGELOG.md
@@ -1,3 +1,13 @@
-# Change Log
-## 1.0.0 - 12/28/2021
+# Changelog
+## [1.0.0] - 12/28/2021
+### Changed
 * First stable release :)
+## [1.1.0] - 12/30/2021
+### Added
+* BPM detection
+* Internal clock
+* Fractional clock values
+* Set BPM via menu
+* Set trim via menu for cv in, arp out, and bass out 
+### Changed
+* Split theory and gui out into separate libs